lesson1/requirements.txt
# Core Dependencies
anthropic>=0.40.0
claude-agent-sdk>=0.1.9  # Latest version as of Nov 2025
requests>=2.31.0
python-dotenv>=1.0.0

# For testing and comparison
colorama>=0.4.6
tabulate>=0.9.0
